fix: Separate MappingsTableContent to a new file
